+++ integration-tests/api/__tests__/admin/draft-order.js
@@ -71,6 +71,52 @@ describe("/admin/draft-orders", () => {
       expect(response.status).toEqual(200)
     })
 
+    it("creates a draft order with a custom shipping option price", async () => {
+      const api = useApi()
+
+      const payload = {
+        email: "oli@test.dk",
+        shipping_address: "oli-shipping",
+        items: [
+          {
+            variant_id: "test-variant",
+            quantity: 2,
+            metadata: {},
+          },
+        ],
+        region_id: "test-region",
+        customer_id: "oli-test",
+        shipping_methods: [
+          {
+            option_id: "test-option",
+            price: 500,
+          },
+        ],
+      }
+
+      const response = await api.post("/admin/draft-orders", payload, {
+        headers: {
+          Authorization: "Bearer test_token",
+        },
+      })
+      expect(response.status).toEqual(200)
+
+      const draftOrderId = response.data.draft_order.id
+
+      const draftOrderResponse = await api.get(
+        `/admin/draft-orders/${draftOrderId}`,
+        {
+          headers: {
+            Authorization: "Bearer test_token",
+          },
+        }
+      )
+      expect(draftOrderResponse.status).toEqual(200)
+      expect(draftOrderResponse.data.draft_order.cart.shipping_total).toEqual(
+        500
+      )
+    })
+
     it("creates a draft order with a billing address that is an AddressPayload and a shipping address that is an ID", async () => {
       const api = useApi()
 
@@ -736,6 +782,40 @@ describe("/admin/draft-orders", () => {
 
       expect(response.data.draft_orders).toEqual([])
       expect(response.data.count).toEqual(0)
+    })
+  })
+
+  describe("GET /admin/draft-orders/:id", () => {
+    beforeEach(async () => {
+      await adminSeeder(dbConnection)
+      await draftOrderSeeder(dbConnection)
+    })
+
+    afterEach(async () => {
+      const db = useDb()
+      await db.teardown()
+    })
+
+    it("retrieves a draft-order should include the items totals", async () => {
+      const api = useApi()
+
+      const order = await api.get("/admin/draft-orders/test-draft-order", {
+        headers: {
+          authorization: "Bearer test_token",
+        },
+      })
+
+      expect(order.status).toEqual(200)
+      expect(order.data.draft_order).toEqual(
+        expect.objectContaining({
+          id: "test-draft-order",
+        })
+      )
+
+      order.data.draft_order.cart.items.forEach((item) => {
+        expect(item.total).toBeDefined()
+        expect(item.subtotal).toBeDefined()
+      })
     })
   })