+++ integration-tests/api/__tests__/batch-jobs/api.js
@@ -306,7 +306,6 @@ describe("/admin/batch-jobs", () => {
       await api
         .post(`/admin/batch-jobs/${jobId}/cancel`, {}, adminReqConfig)
         .catch((err) => {
-          console.log(err)
           expect(err.response.status).toEqual(400)
           expect(err.response.data.type).toEqual("not_allowed")
           expect(err.response.data.message).toEqual(