+++ packages/medusa/src/api/routes/admin/products/list-types.ts
@@ -2,6 +2,7 @@ import { ProductService } from "../../../../services"
 
 /**
  * @oas [get] /products/types
+ * deprecated: true
  * operationId: "GetProductsTypes"
  * summary: "List Product Types"
  * description: "Retrieves a list of Product Types."